requirements.txt
annotated-types==0.7.0
anyio==3.7.1
beautifulsoup4==4.12.3
Brotli==1.1.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
dacite==1.8.1
Deprecated==1.2.15
fastapi==0.103.2
ffmpeg-python==0.2.0
future==1.0.0
h11==0.14.0
idna==3.10
jaconv==0.4.0
markdown-it-py==3.0.0
mdurl==0.1.2
mutagen==1.47.0
platformdirs==4.3.6
pycryptodomex==3.21.0
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
pykakasi==2.3.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytube==15.0.0
RapidFuzz==3.11.0
redis==5.2.1
requests==2.32.3
rich==13.9.4
six==1.17.0
sniffio==1.3.1
soundcloud-v2==1.6.0
soupsieve==2.6
spotdl==4.2.10
spotipy==2.25.0
starlette==0.27.0
syncedlyrics==1.0.1
text-unidecode==1.3
typing_extensions==4.12.2
Unidecode==1.3.8
urllib3==2.3.0
uvicorn==0.23.2
websockets==14.1
wrapt==1.17.0
yt-dlp==2024.12.23
ytmusicapi==1.9.1
